Refactor Dashboard page to use UpcomingShifts component
